+++ tests/testthat/test_print.R
@@ -91,12 +91,12 @@ test_that("Print function for PF objects gives previous results", {
   pf_fit <- suppressWarnings(PF_EM(
     Surv(time, status == 2) ~ ph.ecog + age,
     data = .lung, by = 50, id = 1:nrow(.lung),
-    Q_0 = diag(1, 3), Q = diag(1, 3),
+    Q_0 = diag(1, 3), Q = diag(.1, 3),
     max_T = 800,
     control = list(
-      N_fw_n_bw = 50,
-      N_first = 100,
-      N_smooth = 100,
+      N_fw_n_bw = 200,
+      N_first = 200,
+      N_smooth = 200,
       n_max = 5,
       n_threads = 1)))